refactor(test): split test_inspector.cpp by concern (P11-5, #2647)
test_inspector.cpp was 3,126 lines — over the 2,000-line long-tail target.
Carve three concern-scoped sibling TUs out verbatim, each its own executable
mirroring pulp-test-inspector's registration (same libs + PULP_INSPECTOR_NO_LAUNCH
env guard):

- test_inspector_field_edit.cpp — Phase 3b live-editable box-model field edit
  (20 cases; carries its own Phase3bScene/make_key helper namespace).
- test_inspector_eyedropper.cpp — Phase 3c color eyedropper (13 cases; own
  Phase3cScene/make_move/make_click/make_key helpers).
- test_inspector_domains.cpp — JSON-RPC DomainHandler + Console/Audio/State/
  Performance/LiveConstant inspectors (23 cases).

Parent: 3,126 -> 1,859 lines (54 cases). Mirrors the inspector_overlay_*.cpp
source split (#2555). Verified identical coverage: the four executables sum to
572 assertions / 110 cases, byte-for-byte the pre-split pulp-test-inspector
total (277+92+52+151 = 572; 54+20+13+23 = 110). Built + run GPU-enabled on
arm64 with the pinned Skia/Dawn cache.
