+++ test/test_inspector.cpp
@@ -1365,396 +1365,6 @@ TEST_CASE("InspectorOverlay Phase 2.5: panel icon clicks are ignored when "
     REQUIRE(scene.store.locked_anchors().empty());
 }
 
-// ── Phase 3b — live-editable box-model fields ─────────────────────────────
-//
-// Click a numeric value in the property panel → enter edit mode →
-// type / arrow-nudge / Enter to commit (tweak emitted) / Esc to
-// cancel. The tweak path reuses Phase 0b PR-C-1's
-// emit_tweak_for_selection() so persistence (Phase 1 disk write)
-// gets keyboard edits for free.
-//
-// Test surface uses begin_field_edit() / handle_key_event() directly
-// to avoid having to know the panel's exact pixel layout (it's
-// recomputed each paint pass based on root bounds + panel_width_).
-// The click-into-edit path is exercised by an integration test at
-// the bottom that paints first to populate editable_fields_, then
-// clicks at a known panel y-offset.
-
-namespace {
-
-// Build a small root + child with an anchor for tweak-emission tests.
-struct Phase3bScene {
-    View root;
-    View* child = nullptr;
-    TweakStore store;
-    InspectorOverlay overlay{root};
-
-    Phase3bScene() {
-        root.set_bounds({0, 0, 500, 400});
-        auto c = std::make_unique<View>();
-        c->set_anchor_id("figma:3b:1");
-        c->set_bounds({10, 10, 80, 40});
-        c->flex().padding = 8;
-        c->flex().margin = 4;
-        c->flex().preferred_width = 80;
-        c->flex().preferred_height = 40;
-        child = c.get();
-        root.add_child(std::move(c));
-
-        overlay.set_active(true);
-        overlay.set_tweak_store(&store);
-
-        // Select the child via the normal click path.
-        MouseEvent click;
-        click.position = {20, 20};
-        click.is_down = true;
-        overlay.handle_mouse_event(click);
-    }
-};
-
-KeyEvent make_key(KeyCode k, bool is_down = true, uint16_t mods = 0) {
-    KeyEvent e;
-    e.key = k;
-    e.is_down = is_down;
-    e.modifiers = mods;
-    return e;
-}
-
-} // namespace
-
-TEST_CASE("InspectorOverlay Phase 3b: begin_field_edit sets editing state",
-          "[inspect][overlay][phase3b]") {
-    Phase3bScene s;
-    REQUIRE(s.overlay.selected_view() == s.child);
-    REQUIRE_FALSE(s.overlay.is_editing());
-
-    bool ok = s.overlay.begin_field_edit("layout.padding", 8.0f);
-    REQUIRE(ok);
-    REQUIRE(s.overlay.is_editing());
-    REQUIRE(s.overlay.editing_field() == "layout.padding");
-    REQUIRE(s.overlay.edit_buffer() == "8");
-    REQUIRE(s.overlay.edit_caret_pos() == 1);  // caret at end of "8"
-}
-
-TEST_CASE("InspectorOverlay Phase 3b: begin_field_edit refuses without selection",
-          "[inspect][overlay][phase3b]") {
-    View root;
-    InspectorOverlay overlay(root);
-    overlay.set_active(true);
-    // No selection set.
-    bool ok = overlay.begin_field_edit("layout.padding", 8.0f);
-    REQUIRE_FALSE(ok);
-    REQUIRE_FALSE(overlay.is_editing());
-}
-
-TEST_CASE("InspectorOverlay Phase 3b: typing digits extends edit buffer",
-          "[inspect][overlay][phase3b]") {
-    Phase3bScene s;
-    REQUIRE(s.overlay.begin_field_edit("layout.padding", 8.0f));
-    REQUIRE(s.overlay.edit_buffer() == "8");
-
-    // Type "5" → buffer becomes "85"
-    REQUIRE(s.overlay.handle_key_event(make_key(KeyCode::num5)));
-    REQUIRE(s.overlay.edit_buffer() == "85");
-    REQUIRE(s.overlay.edit_caret_pos() == 2);
-
-    // Type "0" → "850"
-    REQUIRE(s.overlay.handle_key_event(make_key(KeyCode::num0)));
-    REQUIRE(s.overlay.edit_buffer() == "850");
-
-    // Real-time preview: View padding should already reflect 850.
-    REQUIRE(s.child->flex().padding == 850.0f);
-}
-
-TEST_CASE("InspectorOverlay Phase 3b: backspace trims buffer",
-          "[inspect][overlay][phase3b]") {
-    Phase3bScene s;
-    s.overlay.begin_field_edit("layout.padding", 8.0f);
-    s.overlay.handle_key_event(make_key(KeyCode::num5));
-    REQUIRE(s.overlay.edit_buffer() == "85");
-
-    REQUIRE(s.overlay.handle_key_event(make_key(KeyCode::backspace)));
-    REQUIRE(s.overlay.edit_buffer() == "8");
-    REQUIRE(s.overlay.edit_caret_pos() == 1);
-}
-
-TEST_CASE("InspectorOverlay Phase 3b: Enter commits tweak to TweakStore",
-          "[inspect][overlay][phase3b]") {
-    Phase3bScene s;
-    s.overlay.begin_field_edit("layout.padding", 8.0f);
-    // Type "5" → buffer is "85"
-    s.overlay.handle_key_event(make_key(KeyCode::num5));
-
-    REQUIRE(s.store.count() == 0);
-    REQUIRE(s.overlay.handle_key_event(make_key(KeyCode::enter)));
-
-    // Edit mode exited, tweak persisted, View reflects committed value.
-    REQUIRE_FALSE(s.overlay.is_editing());
-    REQUIRE(s.store.count() == 1);
-    auto v = s.store.lookup("figma:3b:1", "layout.padding");
-    REQUIRE(v.has_value());
-    // Value stored as float32 — round-trip equality at integer values
-    // is safe.
-    REQUIRE(v->getFloat32() == 85.0f);
-    REQUIRE(s.child->flex().padding == 85.0f);
-}
-
-TEST_CASE("InspectorOverlay Phase 3b: Esc cancels and reverts",
-          "[inspect][overlay][phase3b]") {
-    Phase3bScene s;
-    s.overlay.begin_field_edit("layout.padding", 8.0f);
-    // Bump to "85" — real-time preview mutates the View.
-    s.overlay.handle_key_event(make_key(KeyCode::num5));
-    REQUIRE(s.child->flex().padding == 85.0f);
-
-    REQUIRE(s.overlay.handle_key_event(make_key(KeyCode::escape)));
-    REQUIRE_FALSE(s.overlay.is_editing());
-    REQUIRE(s.store.count() == 0);                 // no tweak emitted
-    REQUIRE(s.child->flex().padding == 8.0f);      // reverted
-}
-
-TEST_CASE("InspectorOverlay Phase 3b: Up arrow increments by 1",
-          "[inspect][overlay][phase3b]") {
-    Phase3bScene s;
-    s.overlay.begin_field_edit("layout.padding", 8.0f);
-
-    REQUIRE(s.overlay.handle_key_event(make_key(KeyCode::up)));
-    REQUIRE(s.overlay.edit_buffer() == "9");
-    REQUIRE(s.child->flex().padding == 9.0f);
-
-    REQUIRE(s.overlay.handle_key_event(make_key(KeyCode::up)));
-    REQUIRE(s.overlay.edit_buffer() == "10");
-    REQUIRE(s.child->flex().padding == 10.0f);
-}
-
-TEST_CASE("InspectorOverlay Phase 3b: Down arrow decrements by 1",
-          "[inspect][overlay][phase3b]") {
-    Phase3bScene s;
-    s.overlay.begin_field_edit("layout.padding", 8.0f);
-
-    REQUIRE(s.overlay.handle_key_event(make_key(KeyCode::down)));
-    REQUIRE(s.overlay.edit_buffer() == "7");
-    REQUIRE(s.child->flex().padding == 7.0f);
-}
-
-TEST_CASE("InspectorOverlay Phase 3b: Shift+Up nudges by 10",
-          "[inspect][overlay][phase3b]") {
-    Phase3bScene s;
-    s.overlay.begin_field_edit("layout.padding", 8.0f);
-
-    REQUIRE(s.overlay.handle_key_event(make_key(KeyCode::up, true, kModShift)));
-    REQUIRE(s.overlay.edit_buffer() == "18");
-    REQUIRE(s.child->flex().padding == 18.0f);
-
-    REQUIRE(s.overlay.handle_key_event(make_key(KeyCode::down, true, kModShift)));
-    REQUIRE(s.overlay.edit_buffer() == "8");
-}
-
-TEST_CASE("InspectorOverlay Phase 3b: Cmd+Up nudges by 100 (Figma semantics)",
-          "[inspect][overlay][phase3b]") {
-    Phase3bScene s;
-    s.overlay.begin_field_edit("layout.padding", 8.0f);
-
-#ifdef __APPLE__
-    auto main_mod = kModCmd;
-#else
-    auto main_mod = kModCtrl;
-#endif
-    REQUIRE(s.overlay.handle_key_event(make_key(KeyCode::up, true, main_mod)));
-    REQUIRE(s.overlay.edit_buffer() == "108");
-    REQUIRE(s.child->flex().padding == 108.0f);
-}
-
-TEST_CASE("InspectorOverlay Phase 3b: Tab commits and moves to next field",
-          "[inspect][overlay][phase3b]") {
-    Phase3bScene s;
-    // Paint once so editable_fields_ is populated with the panel's
-    // tab order.
-    pulp::canvas::RecordingCanvas canvas;
-    s.overlay.paint(canvas);
-
-    s.overlay.begin_field_edit("layout.padding", 8.0f);
-    s.overlay.handle_key_event(make_key(KeyCode::num5));  // "85"
-
-    REQUIRE(s.overlay.handle_key_event(make_key(KeyCode::tab)));
-    // The previous field's edit was committed (tweak in store).
-    REQUIRE(s.store.count() == 1);
-    REQUIRE(s.store.lookup("figma:3b:1", "layout.padding").has_value());
-    // We're now editing a different field (whatever follows padding
-    // in the panel's draw order).
-    REQUIRE(s.overlay.is_editing());
-    REQUIRE(s.overlay.editing_field() != "layout.padding");
-}
-
-TEST_CASE("InspectorOverlay Phase 3b: clicking a numeric value enters edit mode",
-          "[inspect][overlay][phase3b]") {
-    Phase3bScene s;
-    // Paint to populate editable_fields_ (the panel-side hit rects).
-    pulp::canvas::RecordingCanvas canvas;
-    s.overlay.paint(canvas);
-
-    // The panel sits at the right edge of root (500). panel_width_
-    // defaults to 300, so panel starts at x = 200. paint_props_section
-    // is called with x = panel_x + 8 = 208, and editable values live
-    // at x + 80 = 288 → click at x = 290 lands on the value column.
-    // The Y coordinate just needs to be inside the props area; the
-    // tree section is the top half (root_h * 0.5 = 200), so click at
-    // y = 210 lands on the first heading. Use the first editable
-    // field's reported rect to pick a guaranteed-correct y.
-    //
-    // Use overlay's public state via paint + then a click that lands
-    // within a row. We don't have direct access to editable_fields_
-    // from tests, so click at a coordinate that the first row covers
-    // given the default panel layout. Tree height is root_h/2 = 200,
-    // so props_y = 200; first prop row starts ~ y = 204 + heading
-    // height. We pick a y of 270 which falls into the layout section
-    // ("padding" row given the order width/height/padding/margin).
-    //
-    // To make this robust regardless of exact row arithmetic, scan a
-    // band of y values clicking at the value column until edit mode
-    // is entered. Bounded by the visible props area.
-    bool entered = false;
-    for (float y = 210.0f; y < 380.0f && !entered; y += 5.0f) {
-        MouseEvent click;
-        click.position = {290.0f, y};
-        click.is_down = true;
-        s.overlay.handle_mouse_event(click);
-        entered = s.overlay.is_editing();
-        if (entered) break;
-        // Repaint after each scan to refresh hit rects.
-        canvas.clear();
-        s.overlay.paint(canvas);
-    }
-    REQUIRE(entered);
-}
-
-TEST_CASE("InspectorOverlay Phase 3b: editing layout.width writes to preferred_width",
-          "[inspect][overlay][phase3b]") {
-    Phase3bScene s;
-    s.overlay.begin_field_edit("layout.width", s.child->flex().preferred_width);
-    REQUIRE(s.overlay.edit_buffer() == "80");
-
-    // Type "5" → buffer becomes "805"
-    s.overlay.handle_key_event(make_key(KeyCode::num5));
-    REQUIRE(s.child->flex().preferred_width == 805.0f);
-
-    s.overlay.handle_key_event(make_key(KeyCode::enter));
-    auto v = s.store.lookup("figma:3b:1", "layout.width");
-    REQUIRE(v.has_value());
-    REQUIRE(v->getFloat32() == 805.0f);
-}
-
-TEST_CASE("InspectorOverlay Phase 3b: editing style.opacity round-trips",
-          "[inspect][overlay][phase3b]") {
-    Phase3bScene s;
-    s.child->set_opacity(0.5f);
-    s.overlay.begin_field_edit("style.opacity", 0.5f);
-
-    // Buffer for 0.5 should be "0.50"
-    REQUIRE(s.overlay.edit_buffer() == "0.50");
-
-    // Nudge up: 0.50 + 1 = 1.50, then clamped to 1.0 by set_opacity.
-    s.overlay.handle_key_event(make_key(KeyCode::up));
-    // Underlying View is clamped.
-    REQUIRE(s.child->opacity() == 1.0f);
-    s.overlay.handle_key_event(make_key(KeyCode::enter));
-
-    auto v = s.store.lookup("figma:3b:1", "style.opacity");
-    REQUIRE(v.has_value());
-}
-
-TEST_CASE("InspectorOverlay Phase 3b: cancel_field_edit on inactive is no-op",
-          "[inspect][overlay][phase3b]") {
-    Phase3bScene s;
-    REQUIRE_FALSE(s.overlay.is_editing());
-    s.overlay.cancel_field_edit();  // must not crash
-    REQUIRE_FALSE(s.overlay.is_editing());
-    REQUIRE(s.store.count() == 0);
-}
-
-TEST_CASE("InspectorOverlay Phase 3b: commit_field_edit with empty buffer "
-          "exits edit without emitting tweak",
-          "[inspect][overlay][phase3b]") {
-    Phase3bScene s;
-    s.overlay.begin_field_edit("layout.padding", 8.0f);
-    // Backspace twice to empty the buffer ("8" → "")
-    s.overlay.handle_key_event(make_key(KeyCode::backspace));
-    REQUIRE(s.overlay.edit_buffer().empty());
-
-    REQUIRE_FALSE(s.overlay.commit_field_edit());
-    REQUIRE_FALSE(s.overlay.is_editing());
-    REQUIRE(s.store.count() == 0);
-}
-
-TEST_CASE("InspectorOverlay Phase 3b: caret moves with Left/Right/Home/End",
-          "[inspect][overlay][phase3b]") {
-    Phase3bScene s;
-    s.overlay.begin_field_edit("layout.padding", 80.0f);
-    REQUIRE(s.overlay.edit_buffer() == "80");
-    REQUIRE(s.overlay.edit_caret_pos() == 2);
-
-    REQUIRE(s.overlay.handle_key_event(make_key(KeyCode::left)));
-    REQUIRE(s.overlay.edit_caret_pos() == 1);
-
-    REQUIRE(s.overlay.handle_key_event(make_key(KeyCode::home)));
-    REQUIRE(s.overlay.edit_caret_pos() == 0);
-
-    REQUIRE(s.overlay.handle_key_event(make_key(KeyCode::end_)));
-    REQUIRE(s.overlay.edit_caret_pos() == 2);
-
-    REQUIRE(s.overlay.handle_key_event(make_key(KeyCode::right)));
-    REQUIRE(s.overlay.edit_caret_pos() == 2);  // already at end
-}
-
-TEST_CASE("InspectorOverlay Phase 3b: deactivating inspector cancels active edit",
-          "[inspect][overlay][phase3b]") {
-    Phase3bScene s;
-    s.overlay.begin_field_edit("layout.padding", 8.0f);
-    s.overlay.handle_key_event(make_key(KeyCode::num5));
-    REQUIRE(s.child->flex().padding == 85.0f);
-    REQUIRE(s.overlay.is_editing());
-
-    s.overlay.set_active(false);
-    REQUIRE_FALSE(s.overlay.is_editing());
-    REQUIRE(s.child->flex().padding == 8.0f);   // reverted via cancel
-    REQUIRE(s.store.count() == 0);              // no tweak emitted
-}
-
-TEST_CASE("InspectorOverlay Phase 3b: Esc while editing does NOT exit inspector",
-          "[inspect][overlay][phase3b]") {
-    // Spec: Esc inside an edit cancels the edit; Esc with no edit
-    // exits the inspector. The two semantics share a key, the order
-    // matters.
-    Phase3bScene s;
-    s.overlay.begin_field_edit("layout.padding", 8.0f);
-    REQUIRE(s.overlay.is_active());
-
-    s.overlay.handle_key_event(make_key(KeyCode::escape));
-    REQUIRE_FALSE(s.overlay.is_editing());
-    REQUIRE(s.overlay.is_active());  // still active
-
-    // Second Esc with no edit → exits the inspector.
-    s.overlay.handle_key_event(make_key(KeyCode::escape));
-    REQUIRE_FALSE(s.overlay.is_active());
-}
-
-TEST_CASE("InspectorOverlay Phase 3b: Tab without paint does nothing extra",
-          "[inspect][overlay][phase3b]") {
-    // Tab consults the editable_fields_ list populated by the last
-    // paint. If we Tab without ever painting, commit happens but no
-    // next-field move occurs (gracefully).
-    Phase3bScene s;
-    s.overlay.begin_field_edit("layout.padding", 8.0f);
-    s.overlay.handle_key_event(make_key(KeyCode::num5));
-
-    REQUIRE(s.overlay.handle_key_event(make_key(KeyCode::tab)));
-    // First-field commit happened.
-    REQUIRE(s.store.count() == 1);
-    // Without a prior paint, editable_fields_ is empty, so Tab can't
-    // find a next field — we're no longer editing.
-    REQUIRE_FALSE(s.overlay.is_editing());
-}
-
 // ── Phase 0b PR-C-2: property panel dot indicators ─────────────────────────
 //
 // When the selected view has tweaks in the TweakStore, the property
@@ -1870,300 +1480,6 @@ TEST_CASE("InspectorOverlay: bypassed tweak suppresses dot",
     pulp::canvas::RecordingCanvas bypassed;
     overlay.paint(bypassed);
     REQUIRE(bypassed.command_count() < with_dot_count);
-}
-
-// ── Phase 3c — color eyedropper ───────────────────────────────────────────
-//
-// Eyedropper mode (E key) samples a color from the rendered UI and
-// applies it as a tweak to the selected view's color property. With
-// no Canvas pixel-readback in headless tests, sampling falls back to
-// the resolved background color of the view under the cursor; the
-// tweak path reuses emit_tweak_for_selection() with source
-// "inspector-eyedropper".
-
-namespace {
-
-// Build a root + child where the child carries an explicit background
-// color (so the resolved-style fallback has something to sample) and
-// an anchor (so the eyedropper tweak can land in the store).
-struct Phase3cScene {
-    View root;
-    View* child = nullptr;
-    TweakStore store;
-    InspectorOverlay overlay{root};
-
-    explicit Phase3cScene(Color child_bg = Color::rgba8(0x33, 0x66, 0xcc)) {
-        root.set_bounds({0, 0, 500, 400});
-        auto c = std::make_unique<View>();
-        c->set_anchor_id("figma:3c:1");
-        c->set_bounds({40, 40, 120, 80});
-        c->set_background_color(child_bg);
-        child = c.get();
-        root.add_child(std::move(c));
-
-        overlay.set_active(true);
-        overlay.set_tweak_store(&store);
-
-        // Select the child via the normal click path.
-        MouseEvent click;
-        click.position = {60, 60};
-        click.is_down = true;
-        overlay.handle_mouse_event(click);
-    }
-};
-
-MouseEvent make_move(float x, float y) {
-    MouseEvent e;
-    e.position = {x, y};
-    e.is_down = false;
-    return e;
-}
-
-MouseEvent make_click(float x, float y) {
-    MouseEvent e;
-    e.position = {x, y};
-    e.is_down = true;
-    return e;
-}
-
-} // namespace
-
-TEST_CASE("InspectorOverlay Phase 3c: eyedropper defaults OFF",
-          "[inspect][overlay][phase3c]") {
-    View root;
-    InspectorOverlay overlay(root);
-    REQUIRE_FALSE(overlay.eyedropper_active());
-    REQUIRE_FALSE(overlay.eyedropper_has_sample());
-    REQUIRE(overlay.eyedropper_target() == "style.background_color");
-
-    overlay.set_eyedropper_active(true);
-    REQUIRE(overlay.eyedropper_active());
-    overlay.toggle_eyedropper();
-    REQUIRE_FALSE(overlay.eyedropper_active());
-}
-
-TEST_CASE("InspectorOverlay Phase 3c: 'E' key toggles eyedropper only when active",
-          "[inspect][overlay][phase3c]") {
-    View root;
-    InspectorOverlay overlay(root);
-
-    // Inactive — E does nothing.
-    REQUIRE_FALSE(overlay.handle_key_event(make_key(KeyCode::e)));
-    REQUIRE_FALSE(overlay.eyedropper_active());
-
-    overlay.set_active(true);
-    // Active — E toggles.
-    REQUIRE(overlay.handle_key_event(make_key(KeyCode::e)));
-    REQUIRE(overlay.eyedropper_active());
-    REQUIRE(overlay.handle_key_event(make_key(KeyCode::e)));
-    REQUIRE_FALSE(overlay.eyedropper_active());
-
-    // E with a modifier is ignored (reserved for chord shortcuts).
-    overlay.handle_key_event(make_key(KeyCode::e, true, kModCmd));
-    REQUIRE_FALSE(overlay.eyedropper_active());
-}
-
-TEST_CASE("InspectorOverlay Phase 3c: mouse-move samples color under cursor",
-          "[inspect][overlay][phase3c]") {
-    Phase3cScene s(Color::rgba8(0x12, 0xab, 0x5f));
-    s.overlay.set_eyedropper_active(true);
-    REQUIRE_FALSE(s.overlay.eyedropper_has_sample());
-
-    // Move over the child — resolved-style fallback samples its bg.
-    s.overlay.handle_mouse_event(make_move(80, 70));
-    REQUIRE(s.overlay.eyedropper_has_sample());
-    auto c = s.overlay.eyedropper_sample();
-    REQUIRE(c.r8() == 0x12);
-    REQUIRE(c.g8() == 0xab);
-    REQUIRE(c.b8() == 0x5f);
-}
-
-TEST_CASE("InspectorOverlay Phase 3c: click captures color and emits tweak",
-          "[inspect][overlay][phase3c]") {
-    Phase3cScene s(Color::rgba8(0xff, 0x80, 0x00));
-    s.overlay.set_eyedropper_active(true);
-
-    // Hover to sample, then click to apply.
-    s.overlay.handle_mouse_event(make_move(90, 80));
-    bool consumed = s.overlay.handle_mouse_event(make_click(90, 80));
-    REQUIRE(consumed);  // eyedropper consumes the pick click
-
-    // A pick is a single deliberate action — mode auto-disables.
-    REQUIRE_FALSE(s.overlay.eyedropper_active());
-
-    // The tweak landed in the store under the selection's anchor.
-    REQUIRE(s.store.count() == 1);
-    auto v = s.store.lookup("figma:3c:1", "style.background_color");
-    REQUIRE(v.has_value());
-    REQUIRE(std::string(v->getString()) == "#ff8000");
-}
-
-TEST_CASE("InspectorOverlay Phase 3c: click without prior move still picks",
-          "[inspect][overlay][phase3c]") {
-    // Resolved-style sampling is synchronous, so a click with no
-    // preceding mouse-move still captures a real color (covers
-    // scripted / fast-click use).
-    Phase3cScene s(Color::rgba8(0x40, 0x40, 0x40));
-    s.overlay.set_eyedropper_active(true);
-
-    bool consumed = s.overlay.handle_mouse_event(make_click(100, 90));
-    REQUIRE(consumed);
-    REQUIRE(s.store.count() == 1);
-    auto v = s.store.lookup("figma:3c:1", "style.background_color");
-    REQUIRE(v.has_value());
-    REQUIRE(std::string(v->getString()) == "#404040");
-}
-
-TEST_CASE("InspectorOverlay Phase 3c: click refreshes stale hover sample",
-          "[inspect][overlay][phase3c][regression]") {
-    Phase3cScene s(Color::rgba8(0xff, 0x00, 0x00));
-
-    auto second = std::make_unique<View>();
-    second->set_anchor_id("figma:3c:2");
-    second->set_bounds({40, 150, 120, 80});
-    second->set_background_color(Color::rgba8(0x00, 0x66, 0xff));
-    s.root.add_child(std::move(second));
-
-    s.overlay.set_eyedropper_active(true);
-
-    // Hover over the selected child first, then click elsewhere. The
-    // committed tweak must use the click location, not the stale hover
-    // swatch sample.
-    s.overlay.handle_mouse_event(make_move(90, 80));
-    auto hover_sample = s.overlay.eyedropper_sample();
-    REQUIRE(hover_sample.r8() == 0xff);
-    REQUIRE(hover_sample.g8() == 0x00);
-    REQUIRE(hover_sample.b8() == 0x00);
-
-    bool consumed = s.overlay.handle_mouse_event(make_click(90, 170));
-    REQUIRE(consumed);
-
-    auto v = s.store.lookup("figma:3c:1", "style.background_color");
-    REQUIRE(v.has_value());
-    REQUIRE(std::string(v->getString()) == "#0066ff");
-}
-
-TEST_CASE("InspectorOverlay Phase 3c: click never commits a stale sample",
-          "[inspect][overlay][phase3c][regression]") {
-    // Codex P1 (#2434): pressing E and clicking on empty canvas must
-    // not write a stale color into the tweak store. A prior sample can
-    // be left behind by an earlier hover (or by paint_eyedropper_cursor
-    // sampling from the default cursor position before any mouse-move).
-    // When the click lands where the resolved-style fallback finds no
-    // background-colored view, the resample fails — and the pick must
-    // no-op rather than commit the stale color from the old position.
-    Phase3cScene s(Color::rgba8(0xff, 0x00, 0x00));
-    s.overlay.set_eyedropper_active(true);
-
-    // Seed a stale sample by hovering over the colored child.
-    s.overlay.handle_mouse_event(make_move(90, 80));
-    REQUIRE(s.overlay.eyedropper_has_sample());
-    REQUIRE(s.overlay.eyedropper_sample().r8() == 0xff);
-
-    // Click on empty canvas — (60, 300) is inside the root and clear of
-    // the inspector panel (x >= 200), but hits no background-colored
-    // view, so the resolved-style fallback returns false. The stale red
-    // sample must not survive into the store.
-    bool consumed = s.overlay.handle_mouse_event(make_click(60, 300));
-    REQUIRE(consumed);  // the click is still consumed by the eyedropper
-
-    // No tweak — the failed resample invalidated the sample, so
-    // apply_eyedropper_pick() correctly no-ops instead of writing red.
-    REQUIRE(s.store.count() == 0);
-    REQUIRE_FALSE(s.overlay.eyedropper_has_sample());
-}
-
-TEST_CASE("InspectorOverlay Phase 3c: retargeting writes a different property",
-          "[inspect][overlay][phase3c]") {
-    Phase3cScene s(Color::rgba8(0x00, 0x99, 0xff));
-    s.overlay.set_eyedropper_target("style.border_color");
-    s.overlay.set_eyedropper_active(true);
-
-    s.overlay.handle_mouse_event(make_click(80, 70));
-    auto v = s.store.lookup("figma:3c:1", "style.border_color");
-    REQUIRE(v.has_value());
-    REQUIRE(std::string(v->getString()) == "#0099ff");
-
-    // Empty retarget is ignored — keeps the previous path.
-    s.overlay.set_eyedropper_target("");
-    REQUIRE(s.overlay.eyedropper_target() == "style.border_color");
-}
-
-TEST_CASE("InspectorOverlay Phase 3c: sample_color_at hex round-trips alpha",
-          "[inspect][overlay][phase3c]") {
-    // A semi-transparent background encodes as 8-digit "#rrggbbaa".
-    Phase3cScene s(Color::rgba8(0x20, 0x30, 0x40, 0x80));
-    s.overlay.set_eyedropper_active(true);
-    s.overlay.handle_mouse_event(make_click(80, 70));
-    auto v = s.store.lookup("figma:3c:1", "style.background_color");
-    REQUIRE(v.has_value());
-    REQUIRE(std::string(v->getString()) == "#20304080");
-}
-
-TEST_CASE("InspectorOverlay Phase 3c: eyedropper yields to the panel",
-          "[inspect][overlay][phase3c]") {
-    // A click inside the inspector panel must NOT be eaten by the
-    // eyedropper — the user still needs tree / field interaction.
-    Phase3cScene s;
-    s.overlay.set_eyedropper_active(true);
-
-    // Panel occupies the right 300px of the 500px-wide root.
-    MouseEvent panel_click = make_click(450, 50);
-    bool consumed = s.overlay.handle_mouse_event(panel_click);
-    REQUIRE(consumed);  // panel consumes its own clicks
-    // No pick happened — eyedropper still armed, store untouched.
-    REQUIRE(s.overlay.eyedropper_active());
-    REQUIRE(s.store.count() == 0);
-}
-
-TEST_CASE("InspectorOverlay Phase 3c: pick no-ops without a selection anchor",
-          "[inspect][overlay][phase3c]") {
-    // Hand-authored view with no anchor — the eyedropper still
-    // disables (the click was a deliberate spend) but emits nothing.
-    View root;
-    root.set_bounds({0, 0, 500, 400});
-    auto c = std::make_unique<View>();
-    c->set_bounds({40, 40, 120, 80});
-    c->set_background_color(Color::rgba8(0x11, 0x22, 0x33));
-    root.add_child(std::move(c));
-
-    TweakStore store;
-    InspectorOverlay overlay(root);
-    overlay.set_active(true);
-    overlay.set_tweak_store(&store);
-    overlay.handle_mouse_event(make_click(60, 60));  // select
-    overlay.set_eyedropper_active(true);
-
-    overlay.handle_mouse_event(make_click(80, 70));  // pick
-    REQUIRE_FALSE(overlay.eyedropper_active());
-    REQUIRE(store.count() == 0);
-}
-
-TEST_CASE("InspectorOverlay Phase 3c: deactivating inspector clears eyedropper",
-          "[inspect][overlay][phase3c]") {
-    Phase3cScene s;
-    s.overlay.set_eyedropper_active(true);
-    s.overlay.handle_mouse_event(make_move(80, 70));
-    REQUIRE(s.overlay.eyedropper_has_sample());
-
-    s.overlay.set_active(false);
-    REQUIRE_FALSE(s.overlay.eyedropper_active());
-    REQUIRE_FALSE(s.overlay.eyedropper_has_sample());
-}
-
-TEST_CASE("InspectorOverlay Phase 3c: paint draws swatch when armed",
-          "[inspect][overlay][phase3c]") {
-    // The cursor swatch must paint without crashing once a sample
-    // exists; RecordingCanvas has no pixel readback so this exercises
-    // the resolved-style path + the swatch chrome.
-    Phase3cScene s;
-    s.overlay.set_eyedropper_active(true);
-    s.overlay.handle_mouse_event(make_move(80, 70));
-    REQUIRE(s.overlay.eyedropper_has_sample());
-
-    pulp::canvas::RecordingCanvas canvas;
-    s.overlay.paint(canvas);  // must not crash; swatch chrome drawn
-    REQUIRE(s.overlay.eyedropper_active());
 }
 
 // ── Phase 3a: drag handles ────────────────────────────────────────────────
@@ -2541,586 +1857,3 @@ TEST_CASE("InspectorWindow refreshes console performance and state tabs", "[insp
     REQUIRE(has_label_containing(*tabs->child_at(3), "Range: [-60.00, 12.00] step=0.500  Unit: dB"));
 }
 
-// ── ConsoleCapture ──────────────────────────────────────────────────────────
-
-#include <pulp/inspect/console_capture.hpp>
-
-TEST_CASE("ConsoleCapture: captures log entries") {
-    ConsoleCapture capture;
-    auto cb = capture.callback();
-    cb("log", "hello");
-    cb("warn", "caution");
-    cb("error", "fail");
-    auto entries = capture.entries();
-    REQUIRE(entries.size() == 3);
-    REQUIRE(entries[0].level == "log");
-    REQUIRE(entries[0].message == "hello");
-    REQUIRE(entries[2].level == "error");
-}
-
-TEST_CASE("ConsoleCapture: chains previous callback") {
-    std::string captured;
-    auto previous = [&](std::string_view level, std::string_view msg) {
-        captured = std::string(level) + ":" + std::string(msg);
-    };
-    ConsoleCapture capture;
-    auto cb = capture.callback(previous);
-    cb("log", "test");
-    REQUIRE(captured == "log:test");
-    REQUIRE(capture.entries().size() == 1);
-}
-
-TEST_CASE("ConsoleCapture: clear") {
-    ConsoleCapture capture;
-    auto cb = capture.callback();
-    cb("log", "a");
-    cb("log", "b");
-    capture.clear();
-    REQUIRE(capture.entries().empty());
-}
-
-TEST_CASE("ConsoleCapture: retains the newest ring buffer entries", "[inspect][console][issue-641]") {
-    ConsoleCapture capture;
-    auto cb = capture.callback();
-
-    for (int i = 0; i < 205; ++i)
-        cb("log", "entry-" + std::to_string(i));
-
-    auto entries = capture.entries();
-    REQUIRE(entries.size() == 200);
-    REQUIRE(entries.front().message == "entry-5");
-    REQUIRE(entries.back().message == "entry-204");
-
-    capture.clear();
-    REQUIRE(capture.entries().empty());
-}
-
-// ── AudioInspector ──────────────────────────────────────────────────────────
-
-#include <pulp/inspect/audio_inspector.hpp>
-
-TEST_CASE("AudioInspector: config roundtrip") {
-    AudioInspector audio;
-    AudioConfig cfg;
-    cfg.sample_rate = 48000;
-    cfg.buffer_size = 256;
-    cfg.output_channels = 2;
-    audio.set_config(cfg);
-    auto read = audio.config();
-    REQUIRE(read.sample_rate == 48000);
-    REQUIRE(read.buffer_size == 256);
-}
-
-TEST_CASE("AudioInspector: MIDI logging") {
-    AudioInspector audio;
-    audio.log_midi(0x90, 60, 100, "Note On C4");
-    audio.log_midi(0x80, 60, 0, "Note Off C4");
-    auto events = audio.recent_midi();
-    REQUIRE(events.size() == 2);
-    REQUIRE(events[0].status == 0x90);
-    REQUIRE(events[0].description == "Note On C4");
-}
-
-TEST_CASE("AudioInspector: metering gates level snapshots") {
-    AudioInspector audio;
-
-    audio.report_levels({{0.5f, 0.25f}});
-    REQUIRE(audio.latest_levels().empty());
-
-    audio.set_metering_enabled(true);
-    audio.report_levels({{0.8f, 0.4f}, {0.25f, 0.1f}});
-
-    auto levels = audio.latest_levels();
-    REQUIRE(levels.size() == 2);
-    REQUIRE(levels[0].peak == 0.8f);
-    REQUIRE(levels[1].rms == 0.1f);
-}
-
-// ── DomainHandler ───────────────────────────────────────────────────────────
-
-#include <pulp/inspect/domain_handler.hpp>
-#include <pulp/inspect/state_inspector.hpp>
-#include <pulp/state/store.hpp>
-
-TEST_CASE("DomainHandler: unknown domain") {
-    DomainHandler handler;
-    auto resp = handler.handle(make_request(1, "Bogus.method"));
-    REQUIRE(resp.is_error);
-}
-
-TEST_CASE("DomainHandler: rejects malformed dispatch and missing inspect roots", "[inspect][domain][issue-641]") {
-    DomainHandler handler;
-
-    auto invalid = handler.handle(make_request(1, "DOMGetDocument"));
-    REQUIRE(invalid.is_error);
-    REQUIRE(invalid.params_json == "Invalid method: DOMGetDocument");
-
-    auto unknown = handler.handle(make_request(2, "Bogus.method"));
-    REQUIRE(unknown.is_error);
-    REQUIRE(unknown.params_json == "Unknown domain: Bogus");
-
-    auto dom_missing_root = handler.handle(make_request(3, methods::kDOMGetDocument));
-    REQUIRE(dom_missing_root.is_error);
-    REQUIRE(dom_missing_root.params_json == "No root view attached");
-
-    auto css_missing_root = handler.handle(make_request(4, methods::kCSSGetComputedStyle, R"({"id":"root"})"));
-    REQUIRE(css_missing_root.is_error);
-    REQUIRE(css_missing_root.params_json == "No root view attached");
-}
-
-TEST_CASE("DomainHandler: Inspector.getInfo") {
-    View root;
-    DomainHandler handler;
-    handler.set_root_view(&root);
-    auto resp = handler.handle(make_request(1, "Inspector.getInfo"));
-    REQUIRE_FALSE(resp.is_error);
-    REQUIRE(resp.params_json.find("Pulp") != std::string::npos);
-}
-
-TEST_CASE("DomainHandler: DOM.getDocument") {
-    View root;
-    auto child = std::make_unique<View>();
-    child->set_id("child1");
-    root.add_child(std::move(child));
-    DomainHandler handler;
-    handler.set_root_view(&root);
-    auto resp = handler.handle(make_request(1, "DOM.getDocument"));
-    REQUIRE_FALSE(resp.is_error);
-    REQUIRE(resp.params_json.find("child1") != std::string::npos);
-}
-
-TEST_CASE("DomainHandler: DOM and CSS reject malformed params", "[inspect][domain][issue-641]") {
-    View root;
-    root.set_id("root");
-
-    DomainHandler handler;
-    handler.set_root_view(&root);
-
-    auto dom_bad_json = handler.handle(make_request(1, methods::kDOMGetNodeById, "{"));
-    REQUIRE(dom_bad_json.is_error);
-    REQUIRE(dom_bad_json.params_json == "Invalid params for DOM.getNodeById");
-
-    auto dom_missing_id = handler.handle(make_request(2, methods::kDOMGetNodeById, R"({"nodeId":"root"})"));
-    REQUIRE(dom_missing_id.is_error);
-    REQUIRE(dom_missing_id.params_json == "Invalid params for DOM.getNodeById");
-
-    auto search_bad_json = handler.handle(make_request(3, methods::kDOMSearch, "{"));
-    REQUIRE(search_bad_json.is_error);
-    REQUIRE(search_bad_json.params_json == "Invalid params for DOM.search");
-
-    auto css_bad_json = handler.handle(make_request(4, methods::kCSSGetComputedStyle, "{"));
-    REQUIRE(css_bad_json.is_error);
-    REQUIRE(css_bad_json.params_json == "Invalid params for CSS.getComputedStyle");
-
-    auto css_missing_id = handler.handle(make_request(5, methods::kCSSGetComputedStyle, R"({"nodeId":"root"})"));
-    REQUIRE(css_missing_id.is_error);
-    REQUIRE(css_missing_id.params_json == "Invalid params for CSS.getComputedStyle");
-}
-
-TEST_CASE("DomainHandler: State.getParameters") {
-    StateStore store;
-    store.add_parameter({0, "Volume", "dB", {-60.0f, 6.0f, -12.0f}});
-    store.set_value(0, -6.0f);
-    StateInspector state_inspector(store);
-    DomainHandler handler;
-    handler.set_state_inspector(&state_inspector);
-    auto resp = handler.handle(make_request(1, "State.getParameters"));
-    REQUIRE_FALSE(resp.is_error);
-    REQUIRE(resp.params_json.find("Volume") != std::string::npos);
-}
-
-TEST_CASE("DomainHandler: Audio domain exposes config and MIDI log") {
-    AudioInspector audio;
-    AudioConfig cfg;
-    cfg.sample_rate = 48000;
-    cfg.buffer_size = 128;
-    cfg.input_channels = 1;
-    cfg.output_channels = 2;
-    cfg.latency_samples = 64;
-    audio.set_config(cfg);
-    audio.log_midi(0x90, 60, 100, "Note On C4");
-
-    DomainHandler handler;
-    handler.set_audio_inspector(&audio);
-
-    auto config = handler.handle(make_request(1, methods::kAudioGetConfig));
-    REQUIRE_FALSE(config.is_error);
-    REQUIRE(config.params_json.find("\"sample_rate\"") != std::string::npos);
-    REQUIRE(config.params_json.find("48000") != std::string::npos);
-    REQUIRE(config.params_json.find("\"buffer_size\"") != std::string::npos);
-    REQUIRE(config.params_json.find("128") != std::string::npos);
-    REQUIRE(config.params_json.find("\"latency_samples\"") != std::string::npos);
-    REQUIRE(config.params_json.find("64") != std::string::npos);
-
-    REQUIRE_FALSE(audio.metering_enabled());
-    auto metering = handler.handle(make_request(2, methods::kAudioEnableMetering));
-    REQUIRE_FALSE(metering.is_error);
-    REQUIRE(audio.metering_enabled());
-    REQUIRE(metering.params_json.find("\"metering\":true") != std::string::npos);
-
-    auto midi = handler.handle(make_request(3, methods::kAudioGetMidiLog));
-    REQUIRE_FALSE(midi.is_error);
-    REQUIRE(midi.params_json.find("\"status\"") != std::string::npos);
-    REQUIRE(midi.params_json.find("144") != std::string::npos);
-    REQUIRE(midi.params_json.find("\"data1\"") != std::string::npos);
-    REQUIRE(midi.params_json.find("60") != std::string::npos);
-    REQUIRE(midi.params_json.find("\"description\"") != std::string::npos);
-    REQUIRE(midi.params_json.find("Note On C4") != std::string::npos);
-
-    auto unknown = handler.handle(make_request(4, "Audio.unknown"));
-    REQUIRE(unknown.is_error);
-
-    DomainHandler missing_audio;
-    auto missing = missing_audio.handle(make_request(5, methods::kAudioGetConfig));
-    REQUIRE(missing.is_error);
-}
-
-TEST_CASE("DomainHandler: dispatches inspector domain edge paths", "[inspect][domain][issue-641]") {
-    View root;
-    root.set_id("root");
-    root.set_bounds({0, 0, 320, 200});
-    Theme theme;
-    theme.colors["accent.primary"] = Color::rgba8(12, 34, 56);
-    root.set_theme(theme);
-
-    auto child = std::make_unique<Label>("Child");
-    child->set_id("child");
-    child->set_bounds({5, 6, 70, 20});
-    child->set_opacity(0.5f);
-    child->set_visible(false);
-    child->flex().direction = FlexDirection::row;
-    child->flex().gap = 8;
-    child->flex().padding = 3;
-    child->flex().margin = 2;
-    root.add_child(std::move(child));
-
-    InspectorOverlay overlay(root);
-    ConsoleCapture capture;
-    auto console = capture.callback();
-    console("info", "hello");
-
-    StateStore store;
-    ParamInfo gain;
-    gain.id = 9;
-    gain.name = "Gain";
-    gain.unit = "dB";
-    gain.range = {-60.0f, 12.0f, 0.0f, 0.5f};
-    store.add_parameter(gain);
-    StateInspector state(store);
-
-    AudioInspector audio;
-    AudioConfig cfg;
-    cfg.sample_rate = 44100;
-    cfg.buffer_size = 256;
-    audio.set_config(cfg);
-
-    pulp::render::RenderPassManager rpm;
-    rpm.begin_frame();
-    rpm.begin_pass(pulp::render::RenderPassType::content);
-    rpm.end_pass(4.0f, 11);
-    rpm.end_frame();
-
-    DomainHandler handler;
-    handler.set_root_view(&root);
-    handler.set_overlay(&overlay);
-    handler.set_console_capture(&capture);
-    handler.set_state_inspector(&state);
-    handler.set_audio_inspector(&audio);
-    handler.set_render_pass_manager(&rpm);
-
-    auto invalid = handler.handle(make_request(1, "InvalidMethod"));
-    REQUIRE(invalid.is_error);
-    REQUIRE(invalid.params_json.find("Invalid method") != std::string::npos);
-
-    auto enabled = handler.handle(make_request(2, methods::kInspectorEnable));
-    REQUIRE_FALSE(enabled.is_error);
-    REQUIRE(overlay.is_active());
-
-    auto disabled = handler.handle(make_request(3, methods::kInspectorDisable));
-    REQUIRE_FALSE(disabled.is_error);
-    REQUIRE_FALSE(overlay.is_active());
-
-    auto unknown_inspector = handler.handle(make_request(4, "Inspector.nope"));
-    REQUIRE(unknown_inspector.is_error);
-
-    auto node = handler.handle(make_request(5, methods::kDOMGetNodeById, R"({"id":"child"})"));
-    REQUIRE_FALSE(node.is_error);
-    REQUIRE(node.params_json.find("\"id\"") != std::string::npos);
-    REQUIRE(node.params_json.find("child") != std::string::npos);
-    REQUIRE(node.params_json.find("child_count") != std::string::npos);
-
-    auto missing_node = handler.handle(make_request(6, methods::kDOMGetNodeById, R"({"id":"missing"})"));
-    REQUIRE(missing_node.is_error);
-
-    auto bad_node_params = handler.handle(make_request(7, methods::kDOMGetNodeById, "not json"));
-    REQUIRE(bad_node_params.is_error);
-
-    auto highlight = handler.handle(make_request(8, methods::kDOMHighlightNode));
-    REQUIRE_FALSE(highlight.is_error);
-    auto clear = handler.handle(make_request(9, methods::kDOMClearHighlight));
-    REQUIRE_FALSE(clear.is_error);
-
-    auto search = handler.handle(make_request(10, methods::kDOMSearch, R"({"query":"child"})"));
-    REQUIRE_FALSE(search.is_error);
-    REQUIRE(search.params_json.find("child") != std::string::npos);
-
-    auto bad_search = handler.handle(make_request(11, methods::kDOMSearch, "not json"));
-    REQUIRE(bad_search.is_error);
-
-    auto unknown_dom = handler.handle(make_request(12, "DOM.nope"));
-    REQUIRE(unknown_dom.is_error);
-
-    auto style = handler.handle(make_request(13, methods::kCSSGetComputedStyle, R"({"id":"child"})"));
-    REQUIRE_FALSE(style.is_error);
-    REQUIRE(style.params_json.find("direction") != std::string::npos);
-    REQUIRE(style.params_json.find("row") != std::string::npos);
-    REQUIRE(style.params_json.find("visible") != std::string::npos);
-    REQUIRE(style.params_json.find("false") != std::string::npos);
-
-    auto missing_style = handler.handle(make_request(14, methods::kCSSGetComputedStyle, R"({"id":"missing"})"));
-    REQUIRE(missing_style.is_error);
-
-    auto bad_style = handler.handle(make_request(15, methods::kCSSGetComputedStyle, "not json"));
-    REQUIRE(bad_style.is_error);
-
-    auto theme_resp = handler.handle(make_request(16, methods::kCSSGetTheme));
-    REQUIRE_FALSE(theme_resp.is_error);
-    REQUIRE(theme_resp.params_json.find("accent.primary") != std::string::npos);
-
-    auto unknown_css = handler.handle(make_request(17, "CSS.nope"));
-    REQUIRE(unknown_css.is_error);
-
-    auto perf = handler.handle(make_request(18, methods::kPerfGetMetrics));
-    REQUIRE_FALSE(perf.is_error);
-    REQUIRE(perf.params_json.find("total_time_ms") != std::string::npos);
-    REQUIRE(perf.params_json.find("draw_calls") != std::string::npos);
-    REQUIRE(perf.params_json.find("11") != std::string::npos);
-
-    auto tracking = handler.handle(make_request(19, methods::kPerfEnableTracking));
-    REQUIRE_FALSE(tracking.is_error);
-
-    auto unknown_perf = handler.handle(make_request(20, "Performance.nope"));
-    REQUIRE(unknown_perf.is_error);
-
-    auto set_param = handler.handle(make_request(21, methods::kStateSetParameter, R"({"id":9,"value":-12.5})"));
-    REQUIRE_FALSE(set_param.is_error);
-    REQUIRE(store.get_value(9) == -12.5f);
-
-    auto bad_set_param = handler.handle(make_request(22, methods::kStateSetParameter, "not json"));
-    REQUIRE(bad_set_param.is_error);
-
-    auto unknown_state = handler.handle(make_request(23, "State.nope"));
-    REQUIRE(unknown_state.is_error);
-
-    auto console_entries = handler.handle(make_request(24, methods::kConsoleEnable));
-    REQUIRE_FALSE(console_entries.is_error);
-    REQUIRE(console_entries.params_json.find("hello") != std::string::npos);
-
-    auto unknown_console = handler.handle(make_request(25, "Console.nope"));
-    REQUIRE(unknown_console.is_error);
-
-    auto runtime_eval = handler.handle(make_request(26, methods::kRuntimeEvaluate));
-    REQUIRE(runtime_eval.is_error);
-
-    auto hot_reload = handler.handle(make_request(27, methods::kRuntimeGetHotReloadStatus));
-    REQUIRE_FALSE(hot_reload.is_error);
-    REQUIRE(hot_reload.params_json.find("available") != std::string::npos);
-    REQUIRE(hot_reload.params_json.find("false") != std::string::npos);
-
-    auto unknown_runtime = handler.handle(make_request(28, "Runtime.nope"));
-    REQUIRE(unknown_runtime.is_error);
-
-    auto screenshot = handler.handle(make_request(29, methods::kCaptureScreenshot));
-    REQUIRE(screenshot.is_error);
-    auto screenshot_node = handler.handle(make_request(30, methods::kCaptureScreenshotNode));
-    REQUIRE(screenshot_node.is_error);
-    auto unknown_capture = handler.handle(make_request(31, "Capture.nope"));
-    REQUIRE(unknown_capture.is_error);
-
-    DomainHandler missing_sources;
-    REQUIRE(missing_sources.handle(make_request(32, methods::kDOMGetDocument)).is_error);
-    REQUIRE(missing_sources.handle(make_request(33, methods::kCSSGetTheme)).is_error);
-    REQUIRE(missing_sources.handle(make_request(34, methods::kStateGetParameters)).is_error);
-    REQUIRE(missing_sources.handle(make_request(35, methods::kAudioGetConfig)).is_error);
-
-    auto no_console = missing_sources.handle(make_request(36, methods::kConsoleEnable));
-    REQUIRE_FALSE(no_console.is_error);
-    REQUIRE(no_console.params_json == "[]");
-
-    auto no_perf = missing_sources.handle(make_request(37, methods::kPerfGetMetrics));
-    REQUIRE_FALSE(no_perf.is_error);
-    REQUIRE(no_perf.params_json.find("available") != std::string::npos);
-    REQUIRE(no_perf.params_json.find("false") != std::string::npos);
-}
-
-// ─── StateInspector ListenerToken migration (Slice 3) ───────────────────────
-
-TEST_CASE("StateInspector records parameter changes after subscribing",
-          "[inspect][state][listener]") {
-    StateStore store;
-    ParamInfo info;
-    info.id = 42;
-    info.name = "Cutoff";
-    info.unit = "Hz";
-    info.range = {20.0f, 20000.0f, 1000.0f};
-    store.add_parameter(info);
-
-    StateInspector inspector(store);
-
-    REQUIRE(inspector.recent_changes().empty());
-
-    store.set_value(42, 2400.0f);
-    store.set_value(42, 4800.0f);
-
-    auto changes = inspector.recent_changes();
-    REQUIRE(changes.size() == 2);
-    REQUIRE(changes[0].id == 42);
-    REQUIRE(changes[1].value > changes[0].value);
-}
-
-TEST_CASE("Destroying StateInspector removes its listener (no alive-guard)",
-          "[inspect][state][listener]") {
-    StateStore store;
-    ParamInfo info;
-    info.id = 1;
-    info.name = "Gain";
-    info.range = {0.0f, 1.0f, 0.5f};
-    store.add_parameter(info);
-
-    {
-        StateInspector inspector(store);
-        store.set_value(1, 0.25f);
-        REQUIRE(inspector.recent_changes().size() == 1);
-    } // ~StateInspector() — ListenerToken dtor unregisters
-
-    // The store no longer has a live listener pointing at the
-    // destroyed inspector. With the legacy alive-guard pattern, an
-    // entry was still in the listener list (just no-op-checking
-    // alive). With ListenerToken it's actually removed, so this
-    // set_value is a pure atomic store + a notify() that iterates
-    // an empty snapshot. No use-after-free, no leak.
-    store.set_value(1, 0.75f);
-    REQUIRE_THAT(store.get_value(1), Catch::Matchers::WithinAbs(0.75, 0.001));
-}
-
-// ─── Performance.setRepaintFlash (Tier A Slice 6) ───────────────────────────
-
-#include <pulp/render/dirty_tracker.hpp>
-
-TEST_CASE("Performance.setRepaintFlash toggles DirtyTracker::debug_overlay",
-          "[inspect][perf][repaint-flash]") {
-    pulp::render::DirtyTracker dirty;
-    REQUIRE_FALSE(dirty.debug_overlay());
-
-    DomainHandler handler;
-    handler.set_dirty_tracker(&dirty);
-
-    auto enable_req = make_request(1, methods::kPerfSetRepaintFlash,
-                                   R"({"enabled":true})");
-    auto enable_resp = handler.handle(enable_req);
-    REQUIRE_FALSE(enable_resp.is_error);
-    REQUIRE(dirty.debug_overlay());
-
-    auto get_req = make_request(2, methods::kPerfGetRepaintFlash);
-    auto get_resp = handler.handle(get_req);
-    REQUIRE_FALSE(get_resp.is_error);
-    REQUIRE(get_resp.params_json.find("\"enabled\": true")
-            != std::string::npos);
-    REQUIRE(get_resp.params_json.find("\"available\": true")
-            != std::string::npos);
-
-    auto disable_req = make_request(3, methods::kPerfSetRepaintFlash,
-                                    R"({"enabled":false})");
-    auto disable_resp = handler.handle(disable_req);
-    REQUIRE_FALSE(disable_resp.is_error);
-    REQUIRE_FALSE(dirty.debug_overlay());
-}
-
-TEST_CASE("Performance.setRepaintFlash without a tracker reports unavailable",
-          "[inspect][perf][repaint-flash]") {
-    DomainHandler handler;
-    // Deliberately not calling set_dirty_tracker — the inspector
-    // grew the toggle, but the host process may not have wired one
-    // yet. Behavior: get reports available=false; set returns an
-    // error so the UI can grey out the toggle.
-    auto get_resp = handler.handle(
-        make_request(1, methods::kPerfGetRepaintFlash));
-    REQUIRE_FALSE(get_resp.is_error);
-    REQUIRE(get_resp.params_json.find("\"available\": false")
-            != std::string::npos);
-
-    auto set_resp = handler.handle(
-        make_request(2, methods::kPerfSetRepaintFlash,
-                     R"({"enabled":true})"));
-    REQUIRE(set_resp.is_error);
-}
-
-// ─── LiveConstant RPC (Tier A Slice 13) ─────────────────────────────────────
-
-#include <pulp/view/live_constant_editor.hpp>
-
-TEST_CASE("LiveConstant.list returns the registry contents",
-          "[inspect][live-constant]") {
-    auto& registry = pulp::view::LiveConstantRegistry::instance();
-
-    // Seed the registry. PULP_LIVE_CONSTANT macros do this implicitly,
-    // but we call register_constant directly so the test doesn't have
-    // to compile in a TU that already has them.
-    [[maybe_unused]] auto& cutoff =
-        registry.register_constant("test_cutoff", __FILE__, __LINE__,
-                                   /*default*/ 440.0f,
-                                   /*min*/ 20.0f, /*max*/ 20000.0f);
-    [[maybe_unused]] auto& gain =
-        registry.register_constant("test_gain", __FILE__, __LINE__,
-                                   0.0f, -60.0f, 12.0f);
-
-    DomainHandler handler;
-    auto resp = handler.handle(make_request(1, methods::kLiveConstList));
-    REQUIRE_FALSE(resp.is_error);
-    REQUIRE(resp.params_json.find("test_cutoff") != std::string::npos);
-    REQUIRE(resp.params_json.find("test_gain") != std::string::npos);
-    REQUIRE(resp.params_json.find("\"constants\"") != std::string::npos);
-}
-
-TEST_CASE("LiveConstant.set mutates the registry value",
-          "[inspect][live-constant]") {
-    auto& registry = pulp::view::LiveConstantRegistry::instance();
-    [[maybe_unused]] auto& v =
-        registry.register_constant("test_setter", __FILE__, __LINE__,
-                                   1.0f, 0.0f, 10.0f);
-    registry.reset("test_setter");
-
-    DomainHandler handler;
-    auto resp = handler.handle(make_request(
-        1, methods::kLiveConstSet,
-        R"({"name":"test_setter","value":4.5})"));
-    REQUIRE_FALSE(resp.is_error);
-    REQUIRE_THAT(registry.get("test_setter"),
-                 Catch::Matchers::WithinAbs(4.5, 0.001));
-}
-
-TEST_CASE("LiveConstant.set without a name returns an error",
-          "[inspect][live-constant]") {
-    DomainHandler handler;
-    auto resp = handler.handle(make_request(
-        1, methods::kLiveConstSet, R"({"value":1.0})"));
-    REQUIRE(resp.is_error);
-}
-
-TEST_CASE("LiveConstant.reset rolls a value back to its default",
-          "[inspect][live-constant]") {
-    auto& registry = pulp::view::LiveConstantRegistry::instance();
-    [[maybe_unused]] auto& v =
-        registry.register_constant("test_reset", __FILE__, __LINE__,
-                                   2.0f, 0.0f, 10.0f);
-    registry.set("test_reset", 7.5f);
-    REQUIRE_THAT(registry.get("test_reset"),
-                 Catch::Matchers::WithinAbs(7.5, 0.001));
-
-    DomainHandler handler;
-    auto resp = handler.handle(make_request(
-        1, methods::kLiveConstReset, R"({"name":"test_reset"})"));
-    REQUIRE_FALSE(resp.is_error);
-    REQUIRE_THAT(registry.get("test_reset"),
-                 Catch::Matchers::WithinAbs(2.0, 0.001));
-}
